+++ umask.go
@@ -5,12 +5,38 @@ package compat
 
 import (
 	"errors"
+	"sync"
+	"sync/atomic"
 )
 
-const initialUmask = 0o022
+const (
+	initialUmask = 0o022
+	// These are the only bits used by umask.
+	umaskMask = 0o777
+)
 
-var ErrUmaskChanged = errors.New("process umask changed outside compat")
+var (
+	// ErrUmaskChanged reports if the umask was changed outside compat.
+	ErrUmaskChanged = errors.New("umask changed outside compat")
+	// ErrInvalidUmask reports if the UMASK environment variable is > 0o777.
+	ErrInvalidUmask = errors.New("invalid UMASK environment variable")
+
+	savedUmask atomic.Uint64
+	umaskMutex sync.Mutex
+	umaskErr   error
+)
 
 func init() {
 	_ = initUmask()
+}
+
+func validateUmask(umask uint64) {
+	umaskMutex.Lock()
+	defer umaskMutex.Unlock()
+
+	if umask > uint64(umaskMask) {
+		umaskErr = ErrInvalidUmask
+	} else {
+		umaskErr = nil
+	}
 }