+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
+- Add `ErrUmaskChanged` error (Unix-like systems only).
+-	Add `ErrInvalidUmask` error (Windows only).
+
 ### Fixed
 
 ### Changed